Moved commands to their themed folders
