+++ src/Command/Test/TestCurlParserCommand.php
@@ -1,32 +1,31 @@
 <?php
 
-namespace App\Command;
+namespace App\Command\Test;
 
-use App\Service\PantherParser;
-use Facebook\WebDriver\Exception\NoSuchElementException;
-use Facebook\WebDriver\Exception\TimeoutException;
+use App\Service\CurlParser;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
-final class TestPantherParserCommand extends Command
+final class TestCurlParserCommand extends Command
 {
     private const URL_PARAM = 'url';
     private const SEARCH_TAG_PARAM = 'search_tag';
 
     public function __construct(
-        protected PantherParser $pantherParser
-    ) {
+        protected CurlParser $curlParser
+    )
+    {
         parent::__construct();
     }
 
     public function configure()
     {
         $this
-            ->setName('parser:panther:test')
-            ->setDescription('test panther parser')
+            ->setName('parser:curl:test')
+            ->setDescription('test curl parser')
             ->setDefinition(array())
             ->addArgument(
                 self::URL_PARAM,
@@ -42,24 +41,33 @@ final class TestPantherParserCommand extends Command
         ;
     }
 
-    /**
-     * @throws NoSuchElementException
-     * @throws TimeoutException
-     */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
 
         $url = $input->getArgument(self::URL_PARAM);
-        $client = $this->pantherParser->createGetRequest($url);
+        $page = $this->curlParser->getPageContentViaCurl($url);
+        if (!$page) {
+            $io->warning('Cannot open page for parsing.');
+
+            return Command::INVALID;
+        }
 
         $searchTag = $input->getArgument(self::SEARCH_TAG_PARAM);
-        $crawler = $client->waitFor($searchTag);
+        $parseResult = $page->filter($searchTag)->first();
+        if (false === $parseResult) {
+            $io->warning(sprintf(
+                'Page is parsed, tag %s is not found.',
+                $searchTag
+            ));
+
+            return Command::INVALID;
+        }
 
         $io->success(sprintf(
-            'Page is parsed, %s = %s',
+            'Page is parsed, tag found: %s = %s',
             $searchTag,
-            $crawler->filter($searchTag)->first()->html()
+            $parseResult->text()
         ));
 
         return Command::SUCCESS;
